# src/main/java/com/example/bankcards/dto/request/CreateCardRequest.java
package com.example.bankcards.dto.request;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Pattern;
import lombok.Data;
import java.math.BigDecimal;
import java.time.LocalDate;

@Data
public class CreateCardRequest {

    @NotNull(message = "Initial balance is required")
    private BigDecimal initialBalance;

    @Pattern(regexp = "\\d{16}", message = "Card number must be 16 digits")
    private String cardNumber;

    private LocalDate expiryDate;

}